+++ content/admin/partners.py
@@ -6,13 +6,13 @@
 
 from django.contrib import admin
 from django.utils.html import format_html
-from ordered_model.admin import OrderedModelAdmin
 
+from content.base_models import BaseOrderedModelAdmin
 from content.models import Partner
 
 
 @admin.register(Partner)
-class PartnersAdmin(OrderedModelAdmin):
+class PartnersAdmin(BaseOrderedModelAdmin):
     """Настройка отображения списка Partner с предпросмотром логотипа."""
 
     list_display = ('name', 'logo_preview', 'move_up_down_links')
@@ -29,19 +29,6 @@ class PartnersAdmin(OrderedModelAdmin):
             {'fields': ('created_at', 'updated_at'), 'classes': ('collapse',)},
         ),
     )
-
-    def save_model(self, request, obj, form, change):
-        """Сохраняет объект модели в админке.
-
-        При создании нового объекта автоматически перемещает его
-        на верхнюю позицию (в начало списка), чтобы новые элементы
-        отображались первыми. Для уже существующих объектов сохраняет
-        стандартное поведение.
-        """
-        super().save_model(request, obj, form, change)
-        if not change:
-            obj.top()
-            obj.save()
 
     @admin.display(description='Логотип')
     def logo_preview(self, obj):
